+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - CITATION file
 - In `SPNTA.from_pint()`, compute planetary ephemerides if they are absent in the input `TOAs` object.
 - Support for log-spaced frequencies in Fourier GP components
+- Save MCMC samples in VelaFitter
 ## Changed
 - Residual plot in `pyvela-plot` script
 - Made GP noise marginalization the default in `SPNTA`